requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.0
sentence-transformers==2.2.2
chromadb==0.4.18
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai>=1.0.0